+++ deps.ts
@@ -1,3 +1,1 @@
-import { serve } from "https://deno.land/std@0.114.0/http/server.ts";
-
-export { serve };
+export { Application, Router } from "https://deno.land/x/oak/mod.ts";
