+++ netlify-webhook.ts
@@ -1,60 +1,38 @@
-import {
-  buildNetlifyBuildDetails,
-  getDiscordPayload,
-} from "./getDiscordPayload.ts";
-import { parseDeploymentStatus } from "./parseDeploymentStatus.ts";
-import { NetlifyPayload } from "./types.ts";
+import { Application, Router } from "./deps.ts";
+import { logger as defaultLogger } from "./providers/logging/logger.ts";
+import { Logger } from "./providers/logging/interfaces/logger.ts";
+import { deploymentStatusRouteFactory } from "./factories/routes/deployment-status-route-factory.ts";
 
-export function buildHandler() {
-  const discordBotUrl = Deno.env.get("DISCORD_BOT");
+interface ServeOptions {
+  logger?: Logger;
+  port?: number;
+}
 
-  return async function handler(req: Request): Promise<Response> {
-    if (discordBotUrl == null) {
-      return new Response("'DISCORD_BOT' is not defined", { status: 405 });
-    }
+export function serve(options?: ServeOptions): void {
+  const { logger = defaultLogger, port = 8000 } = options ?? {};
 
-    try {
-      const deploymentStatus = parseDeploymentStatus(req);
+  try {
+    const app = new Application();
 
-      switch (req.method) {
-        case "POST": {
-          const netlifyPayload: NetlifyPayload = await req.json();
+    app.addEventListener("error", (event) => {
+      logger.error(event.error);
+    });
 
-          console.log("Webhook called", {
-            deploymentStatus,
-            netlifyPayload: {
-              id: netlifyPayload.id,
-              name: netlifyPayload.name,
-              permalink: netlifyPayload.links?.permalink,
-              detailsLink: buildNetlifyBuildDetails(netlifyPayload),
-            },
-          });
+    const router = new Router();
 
-          const discordPayload = getDiscordPayload(
-            deploymentStatus,
-            netlifyPayload
-          );
+    router.post(
+      ...deploymentStatusRouteFactory({
+        discordBotUrl: Deno.env.get("DISCORD_BOT"),
+        logger,
+      })
+    );
 
-          const request = new Request(discordBotUrl, {
-            method: "POST",
-            body: JSON.stringify(discordPayload),
-            headers: {
-              "content-type": "application/json",
-            },
-          });
+    app.use(router.routes());
 
-          console.log("Called Discord", {
-            discordPayload,
-          });
+    logger.log(`Listening on port ${port}`);
 
-          return await fetch(request);
-        }
-      }
-    } catch (error: unknown) {
-      console.error(error);
-      return new Response("Application Error", { status: 500 });
-    }
-
-    return new Response("Invalid method", { status: 405 });
-  };
+    app.listen({ port });
+  } catch (error) {
+    logger.error(error);
+  }
 }
