+++ providers/netlify/enums/netlify-deployment-status.ts
@@ -1,0 +1,5 @@
+export enum NetlifyDeploymentStatus {
+  Started = "started",
+  Success = "success",
+  Failure = "failure",
+}
