+++ providers/netlify/interfaces/forms.ts
@@ -1,0 +1,6 @@
+import { Storage } from "./storage.ts";
+
+export interface Forms {
+  submissions: Storage;
+  storage: Storage;
+}
