+++ providers/netlify/interfaces/functions.ts
@@ -1,0 +1,6 @@
+import { Storage } from "./storage.ts";
+
+export interface Functions {
+  invocations: Storage;
+  runtime: Storage;
+}
