+++ providers/netlify/interfaces/identity.ts
@@ -1,0 +1,6 @@
+import { Storage } from "./storage.ts";
+
+export interface Identity {
+  active_users: Storage;
+  invite_only_users: Storage;
+}
