+++ providers/netlify/interfaces/links.ts
@@ -1,0 +1,5 @@
+export interface Links {
+  permalink?: string;
+  alias?: string;
+  branch?: null;
+}
