+++ providers/netlify/interfaces/log-access-attributes.ts
@@ -1,0 +1,7 @@
+export interface LogAccessAttributes {
+  type: string;
+  url: string;
+  endpoint: string;
+  path: string;
+  token: string;
+}
