+++ providers/netlify/interfaces/storage.ts
@@ -1,0 +1,5 @@
+export interface Storage {
+  included: number;
+  unit: string;
+  used: number;
+}
